controller 자동차 이름 검증 부분 model로 이동
