+++ src/main/java/racinggame/model/Cars.java
@@ -1,6 +1,7 @@
 package racinggame.model;
 
 import racinggame.commons.Constant;
+import racinggame.exception.InputValueException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -9,6 +10,8 @@ public class Cars {
     List<Car> cars;
 
     public Cars(String inputCarNames) {
+        isValidCarNames(inputCarNames);
+
         List<Car> cars = new ArrayList<>();
         String[] carNames = inputCarNames.split(Constant.COMMA);
 
@@ -23,5 +26,17 @@ public class Cars {
         return cars;
     }
 
+    private void isValidCarNames(String inputCarNames) {
+        String[] carNames = inputCarNames.split(Constant.COMMA);
 
+        if (carNames.length <= 1 || inputCarNames.isEmpty()) {
+            throw new InputValueException(Constant.CAR_NOR_EXIST_ERR_MSG);
+        }
+
+        for (String carName : carNames) {
+            if (carName.length() > Constant.CAR_NAME_LENGTH_LIMIT) {
+                throw new InputValueException(Constant.CAR_NAME_LENGTH_ERR_MSG);
+            }
+        }
+    }
 }
