+++ src/test/java/racinggame/model/CarTest.java
@@ -1,0 +1,2 @@
+package racinggame.model;public class CarTest {
+}
