+++ src/test/java/racinggame/model/CarsTest.java
@@ -1,0 +1,2 @@
+package racinggame.model;public class CarsTest {
+}
